app/requirements.txt
amqp==2.5.2
aniso8601==8.0.0
appdirs==1.4.3
Appium-Python-Client==0.52
APScheduler==3.6.3
asgiref==3.2.7
asn1crypto==1.3.0
attrs==19.3.0
Automat==20.2.0
beautifulsoup4==4.9.0
billiard==3.6.3.0
blinker==1.4
bloom-filter==1.3
boto3==1.13.5
botocore==1.16.5
Brotli==1.0.7
celery==4.4.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
cycler==0.10.0
dicttoxml==1.7.4
Django==3.0.6
dnspython==1.16.0
docker-py==1.10.6
docker-pycreds==0.4.0
docutils==0.15.2
EasyProcess==0.3
eventlet==0.25.2
Flask==1.1.2
Flask-Auth==0.85
Flask-Email==1.4.4
Flask-RESTful==0.3.8
func-timeout==4.3.5
greenlet==0.4.15
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.6.0
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
Jinja2-Redis==0.1.2
jmespath==0.9.5
kafka==1.3.5
kaitaistruct==0.8
kombu==4.6.8
ldap3==2.7
lxml==4.5.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mitmproxy==5.1.1
monotonic==1.5
numpy==1.18.4
opencv-python==4.2.0.34
panda==0.3.1
pandas==1.0.3
parsel==1.6.0
passlib==1.7.2
Pillow==7.1.2
pprint==0.1
Protego==0.1.16
protobuf==3.11.3
publicsuffix2==2.20191221
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pydivert==2.1.0
pyee==7.0.1
PyHamcrest==2.0.2
pymongo==3.10.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.8.0
pyppeteer==0.0.25
pyquery==1.4.1
PySnooper==0.4.0
python-dateutil==2.8.1
python-lambda==11.7.1
pytz==2020.1
PyVirtualDisplay==0.2.5
PyYAML==5.3.1
queuelib==1.5.0
redis==3.5.0
requests==2.23.0
retrying==1.3.3
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
schedule==0.6.0
scrapinghub==2.3.1
Scrapy==2.1.0
selenium==3.141.0
service-identity==18.1.0
six==1.14.0
sortedcontainers==2.1.0
soupsieve==2.0
sqlparse==0.3.1
tornado==6.0.4
tqdm==4.46.0
Twisted==20.3.0
tzlocal==2.1
urllib3==1.25.9
urwid==2.1.0
vine==1.3.0
w3lib==1.21.0
websocket-client==0.57.0
websockets==8.1
Werkzeug==1.0.1
wsproto==0.15.0
wxPython==4.1.0
zipp==3.1.0
zope.interface==5.1.0
zstandard==0.13.0
